feat: enhance API authentication  with admin checks and fix bugs

+++ src/app/api/races/route.ts
@@ -1,8 +1,11 @@
 import { NextRequest, NextResponse } from "next/server";
 import prisma from "@/lib/prisma";
+import {  withRoleAuth, withAdminAuth} from "@/lib/api-auth";
 
 // Obtener todas las carreras
-export async function GET() {
+export async function GET(req: NextRequest) {
+     const { response } = await withRoleAuth(req,["coach","admin","staff"]);
+    if (response) return response;
     const carreras = await prisma.events.findMany({
         orderBy: {
             fecha: "desc",
@@ -13,6 +16,9 @@ export async function GET() {
 
 // Crear una carrera
 export async function POST(req: NextRequest) {
+    const { response } = await withRoleAuth(req,["coach","admin","staff"]);
+    if (response) return response;
+
     const data = await req.json();
     const carrera = await prisma.events.create({ data });
     return NextResponse.json(carrera);
@@ -20,6 +26,9 @@ export async function POST(req: NextRequest) {
 
 // Editar una carrera
 export async function PUT(req: NextRequest) {
+    const { response } = await withAdminAuth(req);
+    if (response) return response;
+
     const data = await req.json();
     const { id, ...rest } = data;
     const carrera = await prisma.events.update({
@@ -31,6 +40,9 @@ export async function PUT(req: NextRequest) {
 
 // Eliminar una carrera
 export async function DELETE(req: NextRequest) {
+    const { response } = await withAdminAuth(req);
+    if (response) return response;
+
     const { id } = await req.json();
 
     await prisma.listado_escuelas.deleteMany({
